Extract filter logic into module
