+++ SRInfraInventorySystemUI/pages/departments/js/department.js
@@ -4,17 +4,18 @@ import DepartmentManager from './department-manager.js';
 
 document.addEventListener('DOMContentLoaded', () => {
     window.departmentManager = new DepartmentManager();
+    window.departmentFilters = window.departmentManager.filterManager;
 });
 
 // Global fonksiyonlar
 export function applyFilters() {
     if (window.departmentManager) {
-        window.departmentManager.applyFilters();
+        window.departmentManager.filterManager.apply();
     }
 }
 
 export function resetFilters() {
     if (window.departmentManager) {
-        window.departmentManager.resetFilters();
+        window.departmentManager.filterManager.reset();
     }
 }
